Created new BaseComponent with properties and abstract methods used for all Blazor components.
