+++ src/D20Tek.BlazorComponent.Core/D20Tek.BlazorComponent.Core.csproj
@@ -4,7 +4,7 @@
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <RootNamespace>D20Tek.BlazorComponent</RootNamespace>
+    <RootNamespace>D20Tek.BlazorComponents</RootNamespace>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
     <Authors>DarthPedro</Authors>
     <Company>d20Tek</Company>
@@ -16,5 +16,9 @@
     <PackageTags>Blazor .NET5 .NET6 Components WebAssembly</PackageTags>
     <PackageReleaseNotes>Initial release of base types and utilities.</PackageReleaseNotes>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Components" Version="6.0.1" />
+  </ItemGroup>
 
 </Project>
