+++ src/D20Tek.BlazorComponents.Spinner/SpinnerBase.cs
@@ -5,29 +5,9 @@ using Microsoft.AspNetCore.Components;
 
 namespace D20Tek.BlazorComponents
 {
-    public abstract class SpinnerBase : ComponentBase
+    public abstract class SpinnerBase : BaseComponent
     {
         [Parameter]
-        public bool IsVisible { get; set; } = true;
-
-        [Parameter]
         public Size Size { get; set; } = Size.Small;
-
-        [Parameter(CaptureUnmatchedValues = true)]
-        public Dictionary<string, object> RemainingAttributes { get; set; } = new Dictionary<string, object>();
-
-        protected string? CssClass { get; set; } = null;
-
-        protected string? CssStyles { get; set; } = null;
-
-        protected override void OnParametersSet()
-        {
-            this.CssClass = this.CalculateCssClasses();
-            this.CssStyles = this.CalculateCssStyles();
-        }
-
-        protected abstract string? CalculateCssClasses();
-
-        protected abstract string? CalculateCssStyles();
     }
 }
